requirements.txt
flake8 
mypy
typing 
pillow
nose2
numpy 